# [core/state] Do not initialize a sparse array with a long `List[str]`