Fixes #2263.

This fixes an issue of #2263. This also fixes another issue https://github.com/oils-for-unix/oils/issues/2263#issuecomment-2677382000. In the previous versions, `unset[-1]=1` has created `unset=(1)` in the Python version and has core-dumped in the C++ version.


```console
$ bash -c 'unset[-1]=1; declare -p unset'
bash: line 1: unset[-1]: bad array subscript
$ bin/osh -c 'unset[-1]=1; declare -p unset'
declare -a unset=(1)
$ osh-0.27.0 -c 'unset[-1]=1; declare -p unset'  # 0.27.0 C++ version
Segmentation fault         (core dumped) osh-0.27.0 -c 'unset[-1]=1; declare -p unset'
```

The last commit 069162fd6 adds a spec test for the above bug.

This PR contains another commit a271b53c61b69413d53e136511131f13ccc4a413 to refactor the creation of a new sparse array. We've been passed an empty `List[str]` to `BashArray_FromList(list)`, but it is unnecessary to generate any list for an empty new sparse array. This commit adds a new function `BashArray_New()`.