@akinomyoga 2025-02-24
| | Python version | C++ version | Bash 5.2 | Bash 5.3-beta |
|---|---|---|---|---|
| Without eval_unsafe_arith | 69.30 MB (9.17s) | 38.23 MB (0.313s) | 31.97 MB (0.364s) | 14.77 MB (0.178s) |
| With eval_unsafe_arith | 69.29 MB (9.28s) | 38.57 MB (0.311s) | N/A | N/A |

Now, both the speed and the memory size are comparable to Bash. The C++ version is faster than Bash 5.2 and twice as slow as Bash 5.3. The memory size is similar to Bash 5.2 and about 2.5x as Bash 5.3.