@andychu 2025-02-24
Oh awesome!

Do we know why bash 5.3 got faster, and uses less memory?  Is it because of some bash malloc changes?