@andychu 2025-02-24
Excellent, thank you!!

Good to see the internal APIs making it easy to handle the error cases