@akinomyoga 2025-02-24
> Do we know why bash 5.3 got faster, and uses less memory? Is it because of some bash malloc changes?

There has been a ridiculous bug in Bash 4.3..5.2, where two copies of a function were generated for every function (and one was not used at all). Actually, I submitted a patch to Bash last June:

https://lists.gnu.org/archive/html/bug-bash/2024-06/msg00000.html