@andychu 2025-02-24
Wow, that's funny!  Great find

(OSH does have benchmarks in the CI, so we would find anything that doubles memory usage that way)

----

Oh well now I am a bit disappointed we're slower than bash 5.3 :-/    There is a GC-optimized build at _bin/cxx-opt+souffle/osh, but I tried it, and it's not much faster (less GC rooting)

YSH can be a lot faster, but I guess Python/mycpp/OSH does a lot of pointer chasing compared to C code, even though there is less copying

(I wonder if "small string optimization" will help ble.sh in particular, since it uses many integer-like strings?   That would save strings of 6 bytes inside the pointer, rather than allocating a GC object)