@andychu 2025-02-24
(Also it's weird that bash doesn't have a good commit history - https://github.com/bminor/bash/commit/ba4ab055 )