@andychu 2025-02-25
> Yeah, but I think what's measured by source ble.sh is just the function generation speed. I believe the execution speed is a different story.

Yeah that's true, for `source` of a file with  many functions,  then the interpreter is doing one thing

1. parsing the function bodies
2. entering it in a dict

And we are ~1x-3x slower on parsing (the 3x is the result of bash 5.2 optimizations I think)

https://oils.pub/release/0.27.0/benchmarks.wwz/osh-parser/

----

So yeah I agree the runtime speed is more important, since we are already at a "reasonable" speed (faster than bash 5.2)

Small string optimization isn't turned on, but I prototyped it here - https://github.com/oils-for-unix/oils/blob/master/mycpp/small_str_test.cc#L35

Although in general I think bigger speed ups will come from the work that @melvinw has done

- GC rooting optimization - this has a big speedup on the "bubble sort" microbenchmark, although I tested `_bin/cxx-opt+souffle/osh` on the SparseArray benchmarks, and it was only a small difference
- some kind of value types optimization -- not sure if/when we will get here, but chasing many pointers is probably expensive
  - this interacts with a few other things, so it's not straightforward - I have a blog post draft that mentions some of them

But it's good that we've fixed this BashArray problem by moving to the sparse representation!